refactor: move sqids.ts to util folder in backend
