+++ backend/src/controllers/timetable/copyTimetable.ts
@@ -5,7 +5,7 @@ import { Section, Timetable, User } from "../../entity/entities.js";
 import { validate } from "../../middleware/zodValidateRequest.js";
 import { timetableRepository } from "../../repositories/timetableRepository.js";
 import { userRepository } from "../../repositories/userRepository.js";
-import sqids, { validSqid } from "../../sqids.js";
+import sqids, { validSqid } from "../../utils/sqids.js";
 import timetableJSON from "../../timetable.json" with { type: "json" };
 
 const dataSchema = z.object({
