+++ backend/src/controllers/timetable/createTimetable.ts
@@ -3,7 +3,7 @@ import { degreeEnum } from "../../../../lib/src/index.js";
 import { Section, Timetable, User } from "../../entity/entities.js";
 import { timetableRepository } from "../../repositories/timetableRepository.js";
 import { userRepository } from "../../repositories/userRepository.js";
-import sqids from "../../sqids.js";
+import sqids from "../../utils/sqids.js";
 import timetableJSON from "../../timetable.json" with { type: "json" };
 
 export const createTimetable = async (req: Request, res: Response) => {
